fix: migration docs page
